lmstudio: move lmstudio functions to a file to not pollute main
